chore(web): 빼먹은 Community Api 쿼리 키 추가
