+++ apps/web/src/pages/community/[subject]/[category]/[id]/index.tsx
@@ -16,6 +16,7 @@ import Posting from 'components/common/posting';
 import { CommunityPostingList } from 'components/community';
 import { ModalProvider } from 'feature/modalContext';
 import { css } from 'styled-components';
+import { prefetchCommunityPost, QUERY_KEYS } from 'utils/api/community';
 
 const CommunityPost: NextPageWithLayout = ({
   subject,
@@ -72,33 +73,24 @@ export const getServerSideProps: GetServerSideProps = async (ctx) => {
   const { subject, category, id } = query as Params;
 
   const queryClient = new QueryClient();
-  let headers = {};
   const boardView = req.cookies['boardView'];
 
-  if (boardView) headers = { ...headers, Cookie: `boardView=${boardView}` };
-  if (session) headers = { ...headers, ACCESS_TOKEN: session.accessToken };
-
-  await queryClient.prefetchQuery(
-    [
-      ...queries.community.all,
+  await queryClient.prefetchQuery({
+    queryKey: [
+      ...QUERY_KEYS.id(id),
       {
         subject,
         category,
-        id,
       },
     ],
-    () =>
-      serverFetcher(
-        `${process.env.NEXT_PUBLIC_SERVER_URL}/supercar/v1/community/${category}/post-id/${id}`,
-        {
-          method: 'GET',
-          headers,
-        }
-      ).then((res) => {
-        const { ok, status, ...rest } = res;
-        return rest;
-      })
-  );
+    queryFn: () =>
+      prefetchCommunityPost({
+        id,
+        category,
+        token: session?.accessToken,
+        boardView,
+      }),
+  });
 
   return {
     props: {
