+++ apps/web/src/utils/api/community/keys.ts
@@ -1,5 +1,5 @@
 export const QUERY_KEYS = {
   all: ['community'] as const,
   community: () => [...QUERY_KEYS.all, 'list'] as const,
-  detail: (id: string) => [...QUERY_KEYS.all, id] as const,
+  id: (id: string) => [...QUERY_KEYS.all, id] as const,
 };
